Raise CoilReadException and CoilReadTimeoutException from read_coil method
read_coil and write_coil to accept floats

+++ nibe/connection/__init__.py
@@ -4,12 +4,12 @@ from nibe.coil import Coil
 
 
 class Connection(ABC):
-    DEFAULT_TIMEOUT = 5
+    DEFAULT_TIMEOUT = 5.0
 
     @abstractmethod
-    async def read_coil(self, coil: Coil, timeout: int = DEFAULT_TIMEOUT) -> Coil:
+    async def read_coil(self, coil: Coil, timeout: float = DEFAULT_TIMEOUT) -> Coil:
         pass
 
     @abstractmethod
-    async def write_coil(self, coil: Coil, timeout: int = DEFAULT_TIMEOUT) -> Coil:
+    async def write_coil(self, coil: Coil, timeout: float = DEFAULT_TIMEOUT) -> Coil:
         pass
